Fix v0.2 CI regressions
- Run clang-format across v0.2 sources (no logic change).
- Replace deprecated std::atomic_load/store<shared_ptr> in search_server with
  plain mutex-guarded access; the lock already serialized the slot, atomic ops
  were redundant and now error under -Wdeprecated-declarations -Werror.
- Add missing <algorithm> include in standard_analyzer_test.cpp (pre-existing
  v0.1 transitive-include bug surfaced on Linux GCC/Clang).
- Cast rng() narrowing in posting_list_bench.cpp:35 (pre-existing v0.1
  -Wconversion -Werror failure).

+++ bench/posting_list_bench.cpp
@@ -32,7 +32,7 @@ std::vector<std::pair<spp::DocId, std::uint32_t>> Sample(std::size_t n,
     std::vector<std::uint32_t> ids;
     ids.reserve(n);
     for (std::size_t i = 0; i < n; ++i) {
-        ids.push_back(rng() % universe);
+        ids.push_back(static_cast<std::uint32_t>(rng() % universe));
     }
     std::sort(ids.begin(), ids.end());
     ids.erase(std::unique(ids.begin(), ids.end()), ids.end());
